+++ packages/agent-browser-extension/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @agent-native/agent-browser-extension
 
+## 0.1.224
+
+### Patch Changes
+
+- Updated dependencies [8fd035c]
+  - @agent-native/core@0.166.1
+
+## 0.1.223
+
+### Patch Changes
+
+- Updated dependencies [c50b009]
+  - @agent-native/core@0.166.0
+
+## 0.1.222
+
+### Patch Changes
+
+- Updated dependencies [8d56ed2]
+  - @agent-native/core@0.165.5
+
 ## 0.1.221
 
 ### Patch Changes